requirements.txt
asn1crypto==0.24.0
bottle==0.12.13
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
future==0.17.0
idna==2.7
pycparser==2.19
PySocks==1.6.8
python-telegram-bot==11.1.0
requests==2.20.0
six==1.11.0
urllib3==1.24
